Fix argument checking
